Create MoveTime
This method is used to move the robot. 
The first parameter, degrees, controls the direction the robot move in. Setting this parameter to 0 will move the robot right, 90 will move the robot straight, 180 will move the robot left, and 270 will move the robot backwards. 
The second parameter, power, controls how fast the robot will move. Setting this parameter to 0 will not move the robot at all. Setting this parameter to 100 will move the robot at its top speed. 
The third parameter, time, controls how long the robot will move in thousandths of a second. 0 will not move the robot at all. 1000 will move the robot for one second.