Create MoveTouch
This method is used to test the MoveTime method. The robot will move in the desired direction with the desired power until the screen is touched. When the screen is touched, it will display the time, in seconds, that the robot moved. This number will have to be multiplied by 1000 to be used as input for the MoveTime method.